// src/interface/interfaces.tsx
export interface ApiErrorResponse {
    detail: string;
  }
  export interface User {
    last_name: string;
    id_number: string;
    first_name: string;
    phone_number: string;
    password: string;
    created_at: string;
    id: number;
    email: string;
    gender: string;
    is_active: boolean;
  }
  